requirements.txt
pytest==7.4.0
requests==2.31.0
loguru==0.7.2
allure-pytest==2.13.2
pyyaml==6.0
pytest-html==4.0.2